+++ my_music_code/lib/Feed/feed_page.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:my_music_code/Feed/Components/feed_horizontal_scroll_component.dart';
 import 'package:my_music_code/Feed/Components/feed_music_grid.dart';
+import 'package:my_music_code/Feed/Components/feed_profile_app_bar.dart';
 import 'package:my_music_code/Feed/Components/new_music_release.dart';
-import 'package:my_music_code/Globals/responsive_container.dart';
-import 'package:my_music_code/Globals/responsive_text.dart';
 import 'package:my_music_code/Globals/style.dart';
 import 'package:my_music_code/Profile/profile_drawer.dart';
 
@@ -19,24 +18,8 @@ class _FeedPageState extends State<FeedPage> {
   Widget build(BuildContext context) {
     return Scaffold(  
       backgroundColor: backgroundColor,
-      appBar: AppBar(
-        backgroundColor: backgroundColor,
-        leading: Builder(builder: (context) {
-          return IconButton(
-            icon: ResponsiveContainer(
-              height: 30,
-              width: 30,
-              borderRadius: BorderRadius.circular(100),
-              isCubic: true,
-              image: DecorationImage(
-                image: NetworkImage(DefaultPlaceholder.image),
-              ),
-            ),
-            onPressed: () => Scaffold.of(context).openDrawer(),
-          );
-        }),
-        title: ResponsiveText(text: "Hi, Milena Baruc, let's play some music.", fontSize: 16),
-      ),
+
+      appBar: feedProfileAppBar(),
       
       drawer: ProfileDrawer(),
       body: SingleChildScrollView(
